cleanup and remove config.h
